requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
aspose-words==21.11.0
autopep8==1.6.0
colorama==0.4.4
Django==4.0
docx2pdf==0.1.8
docxcompose==1.3.4
docxtpl==0.15.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
Pillow==8.4.0
pycodestyle==2.8.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-docx==0.8.11
pytz==2021.3
pywin32==303
reportlab==3.6.5
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.62.3
tzdata==2021.5
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
